Add Source Native unit tests
- shifted CodeSnippetTestCase to testing utils
